+++ src/main/java/game_engine/model/Projectile.java
@@ -17,12 +17,12 @@ public class Projectile extends Entity {
      * @param tangible                  Whether the Projectile is tangible.
      * @param hitBox                    Dimensions of the hit box for the Projectile.
      * @param size                      Size of the Projectile.
-     * @param x                         X-coordinate of the Projectile within the chunk.
-     * @param y                         Y-coordinate of the Projectile within the chunk.
+     * @param x                         X-coordinate of the Projectile.
+     * @param y                         Y-coordinate of the Projectile.
      * @throws NullPointerException     One of the passed arguments is {@code null}.
      * @throws IllegalArgumentException The passed coordinates are invalid.
      */
-    public Projectile(boolean visible, boolean tangible, Dimension hitBox, Dimension size, int x, int y) throws NullPointerException, IllegalArgumentException {
+    public Projectile(boolean visible, boolean tangible, Dimension hitBox, Dimension size, double x, double y) throws NullPointerException, IllegalArgumentException {
         super(visible, tangible, hitBox, size, x, y);
     }
